Move impl Evaluate code to v1_ext/*
